feat: move VotingEndBox inside the display
