+++ pages/voting.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { useRouter } from 'next/router';
-import { Grid, Heading, Center, Flex } from '@chakra-ui/react';
+import { Grid, Center } from '@chakra-ui/react';
 
 import Display from '@packages/components/Display';
 import InputPanel from '@packages/components/InputPanel/InputPanel';
@@ -10,7 +10,7 @@ import type { NextPage } from 'next';
 
 const VotingPage: NextPage = () => {
   const router = useRouter();
-  const { loadParties, isVoting } = useVotingContext();
+  const { loadParties } = useVotingContext();
 
   useEffect(() => {
     const electionId = router.query.electionId;
@@ -28,29 +28,13 @@ const VotingPage: NextPage = () => {
         p="10"
         alignItems="center"
         maxW="1980px"
+        maxH="900px"
       >
-        {isVoting ? <Display /> : <VotingEndBox />}
-        <Flex
-          h="82%"
-          direction="column"
-          justifyContent="center"
-          backgroundColor="#D9D9D9"
-          borderRadius=".625rem"
-        >
-          <InputPanel />
-        </Flex>
+        <Display />
+        <InputPanel />
       </Grid>
     </Center>
   );
 };
-
-function VotingEndBox() {
-  const { endMessage } = useVotingContext();
-  return (
-    <Center w="100%" h="100%">
-      <Heading>{endMessage}</Heading>
-    </Center>
-  );
-}
 
 export default VotingPage;
